business/funding_strategy/requirements.txt
playwright>=1.40.0
pyyaml>=6.0
jinja2>=3.1.0
rich>=13.0.0
httpx>=0.27.0
